Added logout logic, fixed login error
