+++ app/(dashboard)/dashboard/page.tsx
@@ -1,7 +1,11 @@
+'use client'
+import LogOut from "@/components/LogOut";
+
 export default function Dashboard() {
   return (
     <div>
       <h1>Dashboard Page</h1>
+      <LogOut/>
     </div>
   );
 }
