+++ app/layout.tsx
@@ -1,3 +1,4 @@
+'user client'
 import './globals.css';
 import { Inter } from 'next/font/google';
