+++ components/SignUpForm.tsx
@@ -1,18 +1,17 @@
 'use client';
-import React from 'react';
-import signUp from '../app/firebase/auth/signup';
-import handleGoogleSignIn from '../app/firebase/auth/google-ver';
+import {useState, ChangeEvent} from 'react';
+import signUp from '../utils/signup';
 import { useRouter } from 'next/navigation';
-
+import { UserCredential } from 'firebase/auth';
 const SignUpForm: React.FC<{}> = () => {
-  const [email, setEmail] = React.useState('');
-  const [password, setPassword] = React.useState('');
+  const [email, setEmail] = useState<string>('');
+  const [password, setPassword] = useState<string>('');
   const router = useRouter();
 
   const handleForm = async (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
 
-    const { result, error } = await signUp(email, password);
+    const { result, error }: {result: UserCredential | null; error: Error | null} = await signUp(email, password);
 
     if (error) {
       return console.log(error);
@@ -20,14 +19,22 @@ const SignUpForm: React.FC<{}> = () => {
 
     // else successful
     console.log(result);
-    return router.push('/admin');
+    router.push('/login');
+  };
+  
+  const handleEmailChange = (event: ChangeEvent<HTMLInputElement>) => {
+    setEmail(event.target.value);
+  };
+
+  const handlePasswordChange = (event: ChangeEvent<HTMLInputElement>) => {
+    setPassword(event.target.value);
   };
   return (
     <form onSubmit={handleForm} className="form">
       <label htmlFor="email">
         <p>Email</p>
         <input
-          onChange={(e) => setEmail(e.target.value)}
+          onChange={handleEmailChange}
           required
           type="email"
           name="email"
@@ -38,7 +45,7 @@ const SignUpForm: React.FC<{}> = () => {
       <label htmlFor="password">
         <p>Password</p>
         <input
-          onChange={(e) => setPassword(e.target.value)}
+          onChange={handlePasswordChange}
           required
           type="password"
           name="password"
