+++ package.json
@@ -24,6 +24,7 @@
     "eslint": "^8.40.0",
     "eslint-config-next": "13.4.2",
     "firebase": "^9.22.0",
+    "firebase-tools": "^12.1.0",
     "next": "13.4.2",
     "postcss": "8.4.23",
     "react": "18.2.0",
