+++ tsconfig.json
@@ -28,7 +28,7 @@
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts",
-    "app/signup/page.tsx"
+    "app/(auth)/signup/page.tsx"
   ],
   "exclude": ["node_modules"]
 }
